+++ tests/test_app/TestApp/Plugin/TestPlugin/src/Utility/Sitemap.php
@@ -1,0 +1,19 @@
+<?php
+namespace TestPlugin\Utility;
+
+use MeCms\Utility\SitemapBuilder;
+
+class Sitemap extends SitemapBuilder
+{
+    protected function _ignoredMethod()
+    {
+    }
+
+    public function urlMethod1()
+    {
+    }
+
+    public function urlMethod2()
+    {
+    }
+}